+++ app/src/test/java/org/simple/clinic/drugs/selection/custom/CustomDrugEntryEffectHandlerTest.kt
@@ -45,7 +45,8 @@ class CustomDrugEntryEffectHandlerTest {
       { facility },
       uuidGenerator,
       drugFrequencyToLabelMap,
-      uiActions).build()
+      CustomDrugEntryViewEffectHandler(uiActions)::handle
+  ).build()
 
   private val testCase = EffectHandlerTestCase(effectHandler)